requirements.txt
google-generativeai
spacy>=3.0.0
nltk
pandas
matplotlib
seaborn
wordcloud
plotly
streamlit